+++ project/src/components/app/app.tsx
@@ -1,4 +1,5 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { AppRoute } from '../../constants';
 import Layout from '../layout/layout';
 import MainScreen from '../../pages/main-screen/main-screen';
 import LoginScreen from '../../pages/login-screen/login-screen';
@@ -15,18 +16,18 @@ function App({ placeCount }: AppProps): JSX.Element {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<Layout />}>
+        <Route path={AppRoute.Root} element={<Layout />}>
           <Route index element={<MainScreen placeCount={placeCount} />} />
-          <Route path="/login" element={<LoginScreen />} />
+          <Route path={AppRoute.Login} element={<LoginScreen />} />
           <Route
-            path="/favorites"
+            path={AppRoute.Favorites}
             element={
               <PrivateRoute>
                 <FavoritesScreen />
               </PrivateRoute>
             }
           />
-          <Route path="/offer" element={<RoomScreen />} />
+          <Route path={AppRoute.Offer} element={<RoomScreen />} />
           <Route path="*" element={<Page404Screen />} />
         </Route>
       </Routes>
