+++ project/src/components/layout/layout.tsx
@@ -1,6 +1,6 @@
 import { Outlet, useLocation } from 'react-router-dom';
-import Header from '../header/header';
-import Footer from '../footer/footer';
+import Header from './header/header';
+import Footer from './footer/footer';
 import {routes, pageClassNames, hasFooter} from '../../constants';
 
 function Layout(): JSX.Element {
