+++ project/src/components/location/location-item/location-item.tsx
